requirements.txt
openai>=1.0.0
python-dotenv>=1.2.1
nltk>=3.9.2
requests>=2.31.0
beautifulsoup4>=4.12.0
